add like and rate movie

+++ src/main/java/com/example/movie_rating_service/model/Grade.java
@@ -26,5 +26,5 @@ public class Grade {
     private Film film;
 
     @Column(name = "grade_by_user", nullable = false)
-    private int gradeByUser;
+    private double gradeByUser;
 }
